Added a new network option to handle servers with broken PONG responses like Gitter's irc-bridge.
Unlike a proper PONG response

    << PING LAG438605263
    >> :tepper.freenode.net PONG tepper.freenode.net :LAG438605263

Gitter sends

    << PING LAG438514964
    >> PONG :irc.gitter.im

Before this change, HC would not interpret this as a PONG response to the lagcheck, since it's missing the server name at the front and the LAG identifier at the end. With this change, if G=1 is specified for the network, then any PONG response in this format will be considered to be a response to the outstanding lagcheck, if any.